add window to confirm - final!!

+++ employees/src/app/modules/employee/components/employee-main/employee-main.component.ts
@@ -1,10 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { Employee } from '../../models/employee.model';
 import { EmployeeService } from '../../services/employee.service';
-import { Router } from '@angular/router';
+import { NavigationEnd, Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { EmployeeEditComponent } from '../employee-edit/employee-edit.component';
 import { EmployeeAddComponent } from '../employee-add/employee-add.component';
+import { filter } from 'rxjs';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-employee-main',
@@ -14,29 +16,84 @@ import { EmployeeAddComponent } from '../employee-add/employee-add.component';
 export class EmployeeMainComponent implements OnInit {
   employees: Employee[];
 
+  isAddRoute: boolean = false;
+  isEditRoute: boolean = false;
+
   getEmployees(): void {
     this._service.getEmployees({ status: true }).subscribe(e => this.employees = e)
   }
 
   openEditEmployeeDialog(employeeToEdit: Employee): void {
-    this._router.navigate([`/employees/edit/${employeeToEdit.id}`])
+    const dialogRef = this._dialog.open(EmployeeEditComponent, {
+      width: '90%',
+      data: { employee: employeeToEdit },
+      maxHeight: '90vh',
+      disableClose: true
+    });
+  }
+
+  openAddEmployeeDialog(): void {
+    this._dialog.open(EmployeeAddComponent, {
+      width: '90%',
+      maxHeight: '90vh',
+      disableClose: true
+    });
   }
 
   addEmployee(): void {
-    this._router.navigate(['/employees/add'])
+    this.openAddEmployeeDialog()
+  }
+
+  editEmployee(employeeToEdit: Employee): void {
+    this.openEditEmployeeDialog(employeeToEdit)
   }
 
   deleteEmployee(id: number): void {
-    this._service.deleteEmployeeById(id).then(() => this.getEmployees())
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._service.deleteEmployeeById(id).then(() => this.getEmployees())
+        Swal.fire({
+          position: "bottom-end",
+          title: "Deleted!",
+          text: "Employee has been deleted.",
+          icon: "success",
+          showConfirmButton: false,
+          timer: 1000
+        });
+      }
+    });
   }
 
   filterEmployees(value: Employee[]): void {
     this.employees = value;
   }
 
-  constructor(private _service: EmployeeService, private _router: Router, private dialog: MatDialog) { }
+  constructor(private _service: EmployeeService, private _router: Router, private _dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.getEmployees()
+
+    // this._router.events
+    //   .pipe(filter(event => event instanceof NavigationEnd))
+    //   .subscribe((event) => {
+    //     if (typeof (event) == typeof (NavigationEnd)) {
+    //       if ((event as NavigationEnd).url.includes('/add')) {
+    //         console.log('open edit');
+    //         this.openAddEmployeeDialog();
+    //       }
+    //       else if ((event as NavigationEnd).url.includes('/edit')) {
+    //         console.log('open edit');
+    //         this.openEditEmployeeDialog();
+    //       }
+    //     }
+    //   });
   }
 }
